Creating a Headers component and README file
