+++ client/.eslintcache
@@ -1,1 +1,1 @@
-[{"/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/App.js":"1","/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/components/Navbar/Navbars.js":"2","/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/index.js":"3","/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/reportWebVitals.js":"4","/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/components/Documentation.js":"5","/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/components/About.js":"6","/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/components/HomePage/HomePage.js":"7","/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/services/fecthing.js":"8","/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/services/requestApiKey.js":"9","/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/components/HomePage/Explorer.js":"10","/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/components/HomePage/ApiKeyForm.js":"11","/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/hooks/useSearch.js":"12","/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/context/ApiKeyContext.js":"13"},{"size":828,"mtime":1613048706655,"results":"14","hashOfConfig":"15"},{"size":766,"mtime":1613042241750,"results":"16","hashOfConfig":"15"},{"size":637,"mtime":1613042241754,"results":"17","hashOfConfig":"15"},{"size":362,"mtime":1612185621027,"results":"18","hashOfConfig":"15"},{"size":99,"mtime":1613042241750,"results":"19","hashOfConfig":"15"},{"size":96,"mtime":1613042241750,"results":"20","hashOfConfig":"15"},{"size":591,"mtime":1612788557997,"results":"21","hashOfConfig":"15"},{"size":367,"mtime":1613042241758,"results":"22","hashOfConfig":"15"},{"size":400,"mtime":1613047091657,"results":"23","hashOfConfig":"15"},{"size":3693,"mtime":1613048888905,"results":"24","hashOfConfig":"15"},{"size":3493,"mtime":1612788480773,"results":"25","hashOfConfig":"15"},{"size":712,"mtime":1613047072213,"results":"26","hashOfConfig":"15"},{"size":479,"mtime":1613048819620,"results":"27","hashOfConfig":"15"},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1gkvnxm",{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"32"},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"35"},{"filePath":"36","messages":"37","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"38"},{"filePath":"39","messages":"40","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"41","messages":"42","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"35"},{"filePath":"43","messages":"44","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"45","messages":"46","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"47","messages":"48","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"49","messages":"50","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"51","messages":"52","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"53"},{"filePath":"54","messages":"55","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"56","messages":"57","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/App.js",[],"/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/components/Navbar/Navbars.js",[],["58","59"],"/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/index.js",[],["60","61"],"/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/reportWebVitals.js",[],["62","63"],"/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/components/Documentation.js",[],"/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/components/About.js",[],"/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/components/HomePage/HomePage.js",[],"/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/services/fecthing.js",[],"/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/services/requestApiKey.js",[],"/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/components/HomePage/Explorer.js",[],"/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/components/HomePage/ApiKeyForm.js",[],["64","65"],"/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/hooks/useSearch.js",[],"/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/context/ApiKeyContext.js",[],{"ruleId":"66","replacedBy":"67"},{"ruleId":"68","replacedBy":"69"},{"ruleId":"66","replacedBy":"70"},{"ruleId":"68","replacedBy":"71"},{"ruleId":"66","replacedBy":"72"},{"ruleId":"68","replacedBy":"73"},{"ruleId":"66","replacedBy":"74"},{"ruleId":"68","replacedBy":"75"},"no-native-reassign",["76"],"no-negated-in-lhs",["77"],["76"],["77"],["76"],["77"],["76"],["77"],"no-global-assign","no-unsafe-negation"]
+[{"/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/App.js":"1","/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/components/HomePage/HomePage.js":"2","/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/components/Jumbotron/Header.js":"3"},{"size":842,"mtime":1613060426609,"results":"4","hashOfConfig":"5"},{"size":304,"mtime":1613060516597,"results":"6","hashOfConfig":"5"},{"size":472,"mtime":1613061015107,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1gkvnxm",{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/App.js",[],"/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/components/HomePage/HomePage.js",[],"/home/afafibrahimi/Desktop/ReinExercises/theVampireDiares-api/client/src/components/Jumbotron/Header.js",[]]
