+++ client/src/App.js
@@ -2,7 +2,7 @@ import "./App.scss";
 import { Switch, Route } from "react-router-dom";
 import HomePage from "./components/HomePage/HomePage";
 import About from "./components/About";
-import Documentation from "./components/Documentation";
+import Documentation from "./components/Documentation/Documentation";
 import Navbars from "./components/Navbar/Navbars";
 import Container from "react-bootstrap/Container";
 import { ApiKeyProvider } from "./context/ApiKeyContext";
