--- client/src/components/Documentation.js
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function Documentation() {
-  return <div>Hello</div>;
-}
