+++ client/src/components/Documentation/Documentation.js
@@ -1,0 +1,10 @@
+import React from "react";
+import Header from "../Jumbotron/Header";
+
+export default function Documentation() {
+  return (
+    <div>
+      <Header />
+    </div>
+  );
+}
