+++ client/src/components/Jumbotron/Header.scss
@@ -1,0 +1,3 @@
+.text-align {
+  text-align: center;
+}
